+++ src/RcppExports.cpp
@@ -139,21 +139,60 @@ BEGIN_RCPP
 END_RCPP
 }
 // MaximinFrom_cpp
-Rcpp::IntegerVector MaximinFrom_cpp(Rcpp::NumericMatrix d, int n, int first);
-RcppExport SEXP _MaxMin_MaximinFrom_cpp(SEXP dSEXP, SEXP nSEXP, SEXP firstSEXP) {
+Rcpp::IntegerVector MaximinFrom_cpp(Rcpp::NumericMatrix d, int n, int first, int n_threads);
+RcppExport SEXP _MaxMin_MaximinFrom_cpp(SEXP dSEXP, SEXP nSEXP, SEXP firstSEXP, SEXP n_threadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type d(dSEXP);
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
     Rcpp::traits::input_parameter< int >::type first(firstSEXP);
-    rcpp_result_gen = Rcpp::wrap(MaximinFrom_cpp(d, n, first));
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(MaximinFrom_cpp(d, n, first, n_threads));
+    return rcpp_result_gen;
+END_RCPP
+}
+// MaximinMultiFrom_cpp
+Rcpp::List MaximinMultiFrom_cpp(Rcpp::NumericMatrix d, int n, Rcpp::IntegerVector firsts, int n_threads);
+RcppExport SEXP _MaxMin_MaximinMultiFrom_cpp(SEXP dSEXP, SEXP nSEXP, SEXP firstsSEXP, SEXP n_threadsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type d(dSEXP);
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type firsts(firstsSEXP);
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(MaximinMultiFrom_cpp(d, n, firsts, n_threads));
+    return rcpp_result_gen;
+END_RCPP
+}
+// MatrixOffDiagMax_cpp
+Rcpp::NumericVector MatrixOffDiagMax_cpp(Rcpp::NumericMatrix d, int n_threads);
+RcppExport SEXP _MaxMin_MatrixOffDiagMax_cpp(SEXP dSEXP, SEXP n_threadsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type d(dSEXP);
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(MatrixOffDiagMax_cpp(d, n_threads));
+    return rcpp_result_gen;
+END_RCPP
+}
+// RowSqSumsFromMatrix_cpp
+Rcpp::NumericVector RowSqSumsFromMatrix_cpp(Rcpp::NumericMatrix d, int n_threads);
+RcppExport SEXP _MaxMin_RowSqSumsFromMatrix_cpp(SEXP dSEXP, SEXP n_threadsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type d(dSEXP);
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(RowSqSumsFromMatrix_cpp(d, n_threads));
     return rcpp_result_gen;
 END_RCPP
 }
 // MaximinFromPoints_cpp
-Rcpp::IntegerVector MaximinFromPoints_cpp(Rcpp::NumericMatrix points, int n, int first, int mask);
-RcppExport SEXP _MaxMin_MaximinFromPoints_cpp(SEXP pointsSEXP, SEXP nSEXP, SEXP firstSEXP, SEXP maskSEXP) {
+Rcpp::IntegerVector MaximinFromPoints_cpp(Rcpp::NumericMatrix points, int n, int first, int mask, int n_threads);
+RcppExport SEXP _MaxMin_MaximinFromPoints_cpp(SEXP pointsSEXP, SEXP nSEXP, SEXP firstSEXP, SEXP maskSEXP, SEXP n_threadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -161,29 +200,58 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
     Rcpp::traits::input_parameter< int >::type first(firstSEXP);
     Rcpp::traits::input_parameter< int >::type mask(maskSEXP);
-    rcpp_result_gen = Rcpp::wrap(MaximinFromPoints_cpp(points, n, first, mask));
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(MaximinFromPoints_cpp(points, n, first, mask, n_threads));
+    return rcpp_result_gen;
+END_RCPP
+}
+// MaximinMultiFromPoints_cpp
+Rcpp::List MaximinMultiFromPoints_cpp(Rcpp::NumericMatrix points, int n, Rcpp::IntegerVector firsts, int n_threads);
+RcppExport SEXP _MaxMin_MaximinMultiFromPoints_cpp(SEXP pointsSEXP, SEXP nSEXP, SEXP firstsSEXP, SEXP n_threadsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type points(pointsSEXP);
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type firsts(firstsSEXP);
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(MaximinMultiFromPoints_cpp(points, n, firsts, n_threads));
     return rcpp_result_gen;
 END_RCPP
 }
 // RowSumsFromPoints_cpp
-Rcpp::NumericVector RowSumsFromPoints_cpp(Rcpp::NumericMatrix points);
-RcppExport SEXP _MaxMin_RowSumsFromPoints_cpp(SEXP pointsSEXP) {
+Rcpp::NumericVector RowSumsFromPoints_cpp(Rcpp::NumericMatrix points, int n_threads);
+RcppExport SEXP _MaxMin_RowSumsFromPoints_cpp(SEXP pointsSEXP, SEXP n_threadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type points(pointsSEXP);
-    rcpp_result_gen = Rcpp::wrap(RowSumsFromPoints_cpp(points));
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(RowSumsFromPoints_cpp(points, n_threads));
     return rcpp_result_gen;
 END_RCPP
 }
 // RowSqSumsFromPoints_cpp
-Rcpp::NumericVector RowSqSumsFromPoints_cpp(Rcpp::NumericMatrix points);
-RcppExport SEXP _MaxMin_RowSqSumsFromPoints_cpp(SEXP pointsSEXP) {
+Rcpp::NumericVector RowSqSumsFromPoints_cpp(Rcpp::NumericMatrix points, int n_threads);
+RcppExport SEXP _MaxMin_RowSqSumsFromPoints_cpp(SEXP pointsSEXP, SEXP n_threadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type points(pointsSEXP);
-    rcpp_result_gen = Rcpp::wrap(RowSqSumsFromPoints_cpp(points));
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(RowSqSumsFromPoints_cpp(points, n_threads));
+    return rcpp_result_gen;
+END_RCPP
+}
+// RowSumsSqFromPoints_cpp
+Rcpp::List RowSumsSqFromPoints_cpp(Rcpp::NumericMatrix points, int n_threads);
+RcppExport SEXP _MaxMin_RowSumsSqFromPoints_cpp(SEXP pointsSEXP, SEXP n_threadsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type points(pointsSEXP);
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(RowSumsSqFromPoints_cpp(points, n_threads));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -200,13 +268,26 @@ BEGIN_RCPP
 END_RCPP
 }
 // DiameterFromPoints_cpp
-Rcpp::NumericVector DiameterFromPoints_cpp(Rcpp::NumericMatrix points);
-RcppExport SEXP _MaxMin_DiameterFromPoints_cpp(SEXP pointsSEXP) {
+Rcpp::NumericVector DiameterFromPoints_cpp(Rcpp::NumericMatrix points, int n_threads);
+RcppExport SEXP _MaxMin_DiameterFromPoints_cpp(SEXP pointsSEXP, SEXP n_threadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type points(pointsSEXP);
-    rcpp_result_gen = Rcpp::wrap(DiameterFromPoints_cpp(points));
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(DiameterFromPoints_cpp(points, n_threads));
+    return rcpp_result_gen;
+END_RCPP
+}
+// AllFinite_cpp
+bool AllFinite_cpp(SEXP x, int n_threads);
+RcppExport SEXP _MaxMin_AllFinite_cpp(SEXP xSEXP, SEXP n_threadsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type x(xSEXP);
+    Rcpp::traits::input_parameter< int >::type n_threads(n_threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(AllFinite_cpp(x, n_threads));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -221,12 +302,18 @@ static const R_CallMethodDef CallEntries[] = {
     {"_MaxMin_MaxEntropyGreedy_cpp", (DL_FUNC) &_MaxMin_MaxEntropyGreedy_cpp, 3},
     {"_MaxMin_MaxEntropyExact_cpp", (DL_FUNC) &_MaxMin_MaxEntropyExact_cpp, 2},
     {"_MaxMin_MaxEntropyLogDet_cpp", (DL_FUNC) &_MaxMin_MaxEntropyLogDet_cpp, 2},
-    {"_MaxMin_MaximinFrom_cpp", (DL_FUNC) &_MaxMin_MaximinFrom_cpp, 3},
-    {"_MaxMin_MaximinFromPoints_cpp", (DL_FUNC) &_MaxMin_MaximinFromPoints_cpp, 4},
-    {"_MaxMin_RowSumsFromPoints_cpp", (DL_FUNC) &_MaxMin_RowSumsFromPoints_cpp, 1},
-    {"_MaxMin_RowSqSumsFromPoints_cpp", (DL_FUNC) &_MaxMin_RowSqSumsFromPoints_cpp, 1},
+    {"_MaxMin_MaximinFrom_cpp", (DL_FUNC) &_MaxMin_MaximinFrom_cpp, 4},
+    {"_MaxMin_MaximinMultiFrom_cpp", (DL_FUNC) &_MaxMin_MaximinMultiFrom_cpp, 4},
+    {"_MaxMin_MatrixOffDiagMax_cpp", (DL_FUNC) &_MaxMin_MatrixOffDiagMax_cpp, 2},
+    {"_MaxMin_RowSqSumsFromMatrix_cpp", (DL_FUNC) &_MaxMin_RowSqSumsFromMatrix_cpp, 2},
+    {"_MaxMin_MaximinFromPoints_cpp", (DL_FUNC) &_MaxMin_MaximinFromPoints_cpp, 5},
+    {"_MaxMin_MaximinMultiFromPoints_cpp", (DL_FUNC) &_MaxMin_MaximinMultiFromPoints_cpp, 4},
+    {"_MaxMin_RowSumsFromPoints_cpp", (DL_FUNC) &_MaxMin_RowSumsFromPoints_cpp, 2},
+    {"_MaxMin_RowSqSumsFromPoints_cpp", (DL_FUNC) &_MaxMin_RowSqSumsFromPoints_cpp, 2},
+    {"_MaxMin_RowSumsSqFromPoints_cpp", (DL_FUNC) &_MaxMin_RowSumsSqFromPoints_cpp, 2},
     {"_MaxMin_EuclidColFromPoints_cpp", (DL_FUNC) &_MaxMin_EuclidColFromPoints_cpp, 2},
-    {"_MaxMin_DiameterFromPoints_cpp", (DL_FUNC) &_MaxMin_DiameterFromPoints_cpp, 1},
+    {"_MaxMin_DiameterFromPoints_cpp", (DL_FUNC) &_MaxMin_DiameterFromPoints_cpp, 2},
+    {"_MaxMin_AllFinite_cpp", (DL_FUNC) &_MaxMin_AllFinite_cpp, 2},
     {NULL, NULL, 0}
 };