+++ NEWS.md
@@ -1,3 +1,7 @@
+# MaxMin 0.0.0.9007 (development)
+
+- `FarFirst()` parallelized and optimized.
+
 # MaxMin 0.0.0.9006 (development)
 
 - `Grasp()` no longer discards its best-known solution, is ~10.0× faster,